requirements.txt
hikari>=2.0.0.dev108
hikari-lightbulb>=2.2.1
python-dotenv>=0.19.2
uvloop>=0.16; sys_platform != "win32"